chore: replace `check_membership` with `calculate_merkle_root`
